@AlexeyInc 2022-03-12
Comparison code performance in master branch and final solution

name             old time/op    new time/op    delta
GetDomainStat-8     131ms ± 0%       4ms ± 0%   -96.71%

name             old alloc/op   new alloc/op   delta
GetDomainStat-8     133MB ± 0%       0MB ± 0%  -100.00%

name             old allocs/op  new allocs/op  delta
GetDomainStat-8     1.70M ± 0%     0.00M ± 0%  -100.00%